[DOP-8140] Split HDFS and HDFS.Slots to separated submodules
