+++ docs/connection/file_connection/hdfs/connection.rst
@@ -1,0 +1,9 @@
+.. _hdfs-connection:
+
+HDFS connection
+===============
+
+.. currentmodule:: onetl.connection.file_connection.hdfs.connection
+
+.. autoclass:: HDFS
+    :members: get_current, check, path_exists, is_file, is_dir, get_stat, resolve_dir, resolve_file, create_dir, remove_file, remove_dir, rename_dir, rename_file, list_dir, walk, download_file, upload_file
