+++ docs/connection/file_connection/hdfs/index.rst
@@ -1,0 +1,16 @@
+.. _hdfs:
+
+HDFS
+====
+
+.. toctree::
+    :maxdepth: 1
+    :caption: Connection
+
+    connection
+
+.. toctree::
+    :maxdepth: 1
+    :caption: For developers
+
+    slots
