+++ docs/connection/file_connection/hdfs/slots.rst
@@ -1,0 +1,10 @@
+.. _hdfs-slots:
+
+HDFS Slots
+==========
+
+.. currentmodule:: onetl.connection.file_connection.hdfs.slots
+
+.. autoclass:: HDFSSlots
+    :members: normalize_cluster_name, normalize_namenode_host, get_known_clusters, get_cluster_namenodes, get_current_cluster, get_webhdfs_port, is_namenode_active
+    :member-order: bysource
