+++ docs/connection/file_df_connection/spark_hdfs/slots.rst
@@ -6,5 +6,5 @@ Spark HDFS Slots
 .. currentmodule:: onetl.connection.file_df_connection.spark_hdfs.slots
 
 .. autoclass:: SparkHDFSSlots
-    :members: normalize_cluster_name, normalize_namenode_name, get_known_clusters, get_cluster_namenodes, get_current_cluster, get_ipc_port, is_namenode_active
+    :members: normalize_cluster_name, normalize_namenode_host, get_known_clusters, get_cluster_namenodes, get_current_cluster, get_ipc_port, is_namenode_active
     :member-order: bysource
