+++ setup.cfg
@@ -338,7 +338,7 @@ per-file-ignores =
     onetl/connection/file_connection/file_connection.py:
 # WPS220: Found too deep nesting
         WPS220,
-    onetl/connection/file_connection/hdfs.py:
+    onetl/connection/file_connection/hdfs/connection.py:
 # E800  Found commented out code
         E800,
 # F401 'hdfs.ext.kerberos.KerberosClient as CheckForKerberosSupport' imported but unused
